extract each check into a separate module
